Split generation shaders into two subfolders
