+++ src/RealWorld/drawing/TileDrawer.cpp
@@ -8,7 +8,7 @@
 #include <RealWorld/drawing/MinimapLayout.hpp>
 #include <RealWorld/drawing/TileDrawer.hpp>
 #include <RealWorld/drawing/shaders/AllShaders.gen.hpp>
-#include <RealWorld/generation/shaders/generateStructure.glsl.gen.hpp>
+#include <RealWorld/generation/shaders/tiles/generateStructure.glsl.gen.hpp>
 
 using enum vk::DescriptorType;
 using enum vk::ShaderStageFlagBits;
