+++ src/RealWorld/generation/shaders/CMakeLists.txt
@@ -1,19 +1,8 @@
-﻿real_target_sources(RealWorld
+﻿add_subdirectory(tiles)
+add_subdirectory(vegetation)
+
+real_target_sources(RealWorld
     PUBLIC
-        allocBranches.comp          
-        biome.glsl                  
-        consolidateEdges.comp       
-        expandGrass.glsl            
-        expandVegInstances.comp     
-        expansionHelpers.glsl       
-        generateColumnValues.glsl   
-        generateStructure.comp      generateStructure.glsl
         GenerationPC.glsl           
         generationPll.glsl          
-        outputBranches.comp         
-        selectVegSpecies.comp       
-        selectVariant.comp          
-        vegDistribution.glsl        
-        VegPrepSB.glsl              
-        VegTemplatesUB.glsl         
 )
