+++ src/RealWorld/generation/shaders/tiles/CMakeLists.txt
@@ -1,0 +1,10 @@
+﻿real_target_sources(RealWorld
+    PUBLIC
+        biome.glsl                  
+        consolidateEdges.comp       
+        generateColumnValues.glsl   
+        generateStructure.comp      generateStructure.glsl
+        selectVariant.comp          
+        undergroundMaterial.glsl    
+        worleyVoronoiNoise.glsl     
+)
