+++ src/RealWorld/generation/shaders/tiles/consolidateEdges.comp
