+++ src/RealWorld/generation/shaders/vegetation/expandVegInstances.comp
@@ -4,16 +4,16 @@
 #version 460
 #include <RealWorld/constants/Chunk.glsl>
 #include <RealWorld/simulation/general/shaders/random.glsl>
-#include <RealWorld/generation/shaders/generateStructure.glsl>
-#include <RealWorld/generation/shaders/vegDistribution.glsl>
+#include <RealWorld/generation/shaders/tiles/generateStructure.glsl>
+#include <RealWorld/generation/shaders/vegetation/vegDistribution.glsl>
 #include <RealWorld/simulation/vegetation/shaders/normAngles.glsl>
 #include <RealWorld/generation/shaders/generationPll.glsl>
-#include <RealWorld/generation/shaders/VegTemplatesUB.glsl>
+#include <RealWorld/generation/shaders/vegetation/VegTemplatesUB.glsl>
 #define VegPrepSB_ACCESS
-#include <RealWorld/generation/shaders/VegPrepSB.glsl>
+#include <RealWorld/generation/shaders/vegetation/VegPrepSB.glsl>
 #include <RealWorld/simulation/vegetation/shaders/BranchSB.glsl>
 #include <RealWorld/generation/shaders/GenerationPC.glsl>
-#include <RealWorld/generation/shaders/expansionHelpers.glsl>
+#include <RealWorld/generation/shaders/vegetation/expansionHelpers.glsl>
 
 layout (local_size_x = 64,
         local_size_y = 1,
@@ -55,7 +55,7 @@ const int k_maxStackDepth = 8;
 shared TurtleState s_stack[k_maxStackDepth];
 
 // Requires at least some shared variables
-#include <RealWorld/generation/shaders/expandGrass.glsl>
+#include <RealWorld/generation/shaders/vegetation/expandGrass.glsl>
 
 // Return index relative to the provided rules or ~0 if no rule was selected
 uint probSelectRuleIndex(uint cond, RuleBodies rules){
