+++ src/RealWorld/generation/shaders/vegetation/outputBranches.comp
@@ -8,9 +8,9 @@
 #include <RealWorld/simulation/vegetation/shaders/normAngles.glsl>
 #include <RealWorld/generation/shaders/generationPll.glsl>
 #include <RealWorld/simulation/vegetation/shaders/BranchSB.glsl>
-#include <RealWorld/generation/shaders/VegTemplatesUB.glsl>
+#include <RealWorld/generation/shaders/vegetation/VegTemplatesUB.glsl>
 #define VegPrepSB_ACCESS readonly
-#include <RealWorld/generation/shaders/VegPrepSB.glsl>
+#include <RealWorld/generation/shaders/vegetation/VegPrepSB.glsl>
 
 layout (local_size_x = 2,
         local_size_y = 32,
