+++ src/RealWorld/generation/shaders/vegetation/selectVegSpecies.comp
@@ -2,13 +2,13 @@
  *  @author     Dubsky Tomas
  */
 #version 460
-#include <RealWorld/generation/shaders/generateStructure.glsl>
-#include <RealWorld/generation/shaders/vegDistribution.glsl>
+#include <RealWorld/generation/shaders/tiles/generateStructure.glsl>
+#include <RealWorld/generation/shaders/vegetation/vegDistribution.glsl>
 #include <RealWorld/simulation/general/shaders/random.glsl>
 #include <RealWorld/generation/shaders/generationPll.glsl>
 #include <RealWorld/simulation/vegetation/shaders/BranchAllocRegSB.glsl>
 #define VegPrepSB_ACCESS writeonly
-#include <RealWorld/generation/shaders/VegPrepSB.glsl>
+#include <RealWorld/generation/shaders/vegetation/VegPrepSB.glsl>
 #include <RealWorld/generation/shaders/GenerationPC.glsl>
 
 layout (local_size_x = k_maxVegPerChunk,
